move HoleNode from parser package to fixer package
